[od6.0] Bugfix: subpixel jump of preloaded texture during vertical tab-drag.
Summary:
(1) Bugfix: subpixel jump of preloaded texture during vertical tab-drag.
Unveiled because dragger position of random line is not snapped in the
z-direction when using the new tab-dragger instead of the old knot-dragger.
Caused by [a] texture mapping not subpixel accurate in z-direction, and
[b] the getZRange(.) function of RandomSeisDataPack did not tell what it
actually contained under these subpixel circumstances.
(2) The fact that it was possible for the createDataPackFrom(.) function to
call the Regular2RandomDataCopier with incompatible Z-ranges can pretty well
be the reason of its crashes in sysMemCopy(.), which have been reported
without a proper workflow description multiple times (Mantis 4874).      
(3) Better snap the dragger position in the z-direction after tab-drag is
finished. This complies with behaviour of the old knot-dragger as well as 
the tab-dragger of the inl/crl/z-slices. This also gives less ugly z-range
values in the random line positioning dialog.

Test Plan:
Tested in OpendTect with preloaded Seismics texture on default random line.
Texture properties set to Classification for monitoring. The texture no
longer jumps a little bit up and down while dragging one of the vertical
green tabs. Neither does the added programming error on the incompatibility
of z-ranges popup anymore after the code modifications. Also tested with
non-default z-range preloaded. The snapping after finishing tab-drag has
been verified from magnifying the scene and checking values in the random
line positioning dialog.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9310