[od6] Added don't ask again variable to most calls.
Summary:
In many cases, we may want to have the option to ask the user if we
are going to ask a quesiton again. I added a bool* to make it possible to
do so. Kept all API intact. In the od version of this, the API will change.

Test Plan: Tested with issue reporter, and works fine.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D9316