od6.0-critical bug fixing: mouse click position is different from horizon position on screen
Summary:
This bug was fixed in 3d horizon tracking in 2d viewer by Nanne. But it is still there in 2d horizon tracking.
when do mouse click we can see the displayed horizon position has z-shift bias comparing with mouse position.
The solution is same as Nanne does in 3d horizon tracking in 2d viewer.
Since Nanne is in holiday. I send this to you.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9315