od: Added generic IOM().isPresent(name)
Summary: needed where the object name is derived from another name, thus not checked by a uiObject. When it returns false it is safe to create an object with that name, nothing will be overwritten.

Test Plan: tested with two names, one existing, one not. Got twice the correct output.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D9333