[od6.0] Bugfix for inl/crl/z not coinciding with sampling of preloaded attrib
Summary:
Bugfix for inl/crl/z slice not coinciding with sampling of preloaded attrib.
Part of Phabricator T254.
(1) Inl/crl-slices alternately showed seismics and undef color if data
cube has non-default step in scroll direction. Ideally, this would require
interpolation, but instead using the nearest line present in the data
cube is currently preferred for speed. This is totally equivalent to what
already happens when scrolling random lines. We might investigate fast
interpolation in the future, but users not happy with such positional
round-offs should just NOT preload their data with a non-default step
in the scroll direction. Note that limiting the user-selected position
of the plane to only lines with valid data is a cumbersome alternative,
because of the possibility of having multiple preloaded cubes with
different steps in the multi-attribute case.
(2) A slice moved totally outside the preloaded cube range resulted in an
empty datapack instead of one with undef(s). This made that no texture at
all was produced in that area and the slice got a random color not adapting
to the undef color of the current color map. The memory (only temporarily)
gained by this does not outweigh all extra trouble it yields.

Test Plan:
Tested in F3-Demo for inl/crl/z slices and a preloaded seismic cube with
smaller ranges and a bigger step in all directions. No more undef colored
planes in between the seismics when scrolling the slice up and down.
On the other hand, planes do get the correct undef color when moving the
slice totally outside the preloaded cube range. And this color does also
adapt if the user changes the color map, for example.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: nanne, mahant

Maniphest Tasks: T254

Differential Revision: https://phabricator.dgbes.com:8443/D9360