od6.0-improvement of displaying intersection points between 3dhorizon on 3d elements in 2dViewer
Summary:
This revision is ported from D9341 in od trunk.

This is mantis issue 5030 and an extra requirement from Aryat who wants to disappear intersection points if intersection point is
overlapped by a new drawing horizon ( but for 2d horizon, they want to keep the intersection point in this case ).
The solution does:
1.keep intersection point not disappeared when doing mouse click in somewhere.
2.using a TrcKeySsmpling to check if we need to remove intersection point

Test Plan: compiles fine and tests also fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9342