od-added parallel processing in isocontourtracer
Summary: The revision is trying to improve the speed in extracting contour line.

Test Plan:
Compiles fine. 
Tests using single horizon create contour, and create intersection line between single horizon and z-slice, the speed is apparently improved.
For the intersection line computation between multi horizon's and z-slices, the speed in debug version slightly slow than the speed in current
release version. Tested also in the tool of Intel VTune Amplifier, the most of time consuming is in the mIsDef().
So, I would like first to know how is the speed of this revision in the release version.
Later on we will think if the speed is ok, if not we need think an extra revision.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9382