od: Added a function addSpace in uiString
Summary: addSpace is a function in bufferstrings which adds spaces at the end of the string. The function added in uiString takes in number of spaces you want as argument and add those spaces at the end of the string.

Test Plan: Tested visually. Okay button in uitaskrunner displayed Pause as first text and when click the display text changed to Resume. Since space was fixed, Resume was displayed as Resum. This display is now corrected.

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9454