od-bug fixing 5067 about deleting pick in displaying at section
Summary:
This is a bug after delete a pick at the display status of displaying at section, the display on/off array is not updated in osggeo markerset class. 
then the displayed marker of markerset is wrong. in the option of displaying at section, we just display part of markers not all, so the onoff array 
is wrong, the displayed marker also is wrong.

Test Plan: Compiles and Testing fine. The bug is fixed.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9443