od6.0 : Fix for crash in creation of new synthetic
Summary:
If we are creating new synthetic from an existing reflectivity model, then the elasticmodel pointer would be null. Hence if getOutSamplingFromModel() then getTimeSampling() would cause crash as aimodels_ is null. So split up the getOutSamplingFromModel() into two parts. one which chekcs for input & another which gets the sampling. But if getOutSamplingFromModel() fails then it means the reflectivitymodel is screwed. So can we then even generate synthetics ? If not then this fix is fine, if yes then we need to add the getTimeSampling() lines

http://bugs.dgbes.com/view.php?id=4873

Test Plan: Tested with models which create a single elastic layers with no variation & other normal models

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9465