od6.0-bug fixing 5127: Horizon "display at section" in depth science
Summary:
At beginning I made a mistake to load a velocity model with only one inline, I see all objects, horizon, well, 
and so on are not displayed correctly in depth science. So I thought zaxistransform is not working well. It is wrong.

But, the bug still is there when displaying horizon only at section. 
1. in windows debug mode, when loading horizon, it will be crashed because the BinId is undefined.
2. this is bug in the function " TravserLine()" in which the transformed position is skpied by the "if" condition.

This solution fixed above two problems and fixed the bug 5127.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9469