od-bug fixing markerset not displayed caused by revision D9444(committed two days ago).
Summary:
The bug is found by using polygon selection tool selecting fault and fault stick, the selection is not displayed.
The reason is that I added one calling in revision D9444: markerset_->turnMarkerOn( index, draw ); in addPos(0,setPos and insertPos() in vismarkerset.
so if the draw argument is false, the marker will be not displayed.

The revision D9444 is good. but I should do corresponded changes in other codes if the argument "draw" is set as false.
the solution fixed "draw" issue by searching all addPos, setPos, and insertPos in od and dgb.

Test Plan: Compiles and Test fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9479