od6.0: Password in proxy settings dialog is now saved in hex code.
Summary: To address security concerns of some of our clients, Password in proxy settings dialog is now saved in hex code. Also, the old plain text passwords are replaced with the new hex-coded ones.

Test Plan: Did not test with a real proxy server. Just checked the strings in the debugger.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9503