[od6.0] Added support for data/localizations/inprogress
Summary:
Some languages (Portugese & Spanish) should not be in the release, hence they are moved
out of the data/localization/source folder and put in the data/localization/inprogress.

Still the update script needs to find them there and update them in the update process.

Test Plan: Updated run (under dgb6.0 folder), and the in-progress localizations are found and updated correctly.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9507