od6.0-bug fixing 5248
Summary:
The revision is same as revision D9530 in od trunk.
In the report 5248, Farrukh wants that it should not display intersections when 2dline attribute is disable or 2dline has no attributes.
This solution fixed it

Test Plan: Compiles fine. Tests by Farrukh on my machine works as he expected.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9531