od-bug 5220 part fixing about save/restore session of horizon contour
Summary:
This is part fixing of report 5220 ( Another fixing should be in dgb plugin ). The main issue is the par parameters don't have Z attribute for create a contour when restore contour session.
So, the solution added an extra info zDomain in par parameters.

Test Plan: Compiles and Tests fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9534