od: Made the marker default selection same as that is saved in .disp file.
Summary: The code adds all marker names to the list of selected markers which is not needed. Hence removed it. Mantis Issue: http://bugs.dgbes.com/view.php?id=5064.

Test Plan: Only those markers are displayed which are in the selected markers list (in the settings file) after the change.

Reviewers: nageswara

Reviewed By: nageswara

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9561