od-bug 5269 fixing about undo/redo in 2dviewer
Summary:
If there are multi (2d/3d) horizons in the tree, the undo and redo in 2dviewer interpolation always does the times of how many horizons. That is the bug in 5269.
The solution fixed the undo/redo does only current active horizon, not all horizons loaded.

Test Plan: Compiles and Tests fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9569