od--bug fixing 5280 about delete seed from horizon tracking with auto mode
Summary:
This revision is ported from revision D9592 in od6.0

The bug was caused by mouse click on an unavailable position in the searching window range. In this case, it is still added into getPosAttribList in emobject but not as a seed.
So the size of seed array is different from as the size of PosAttribList, when deleting a seed, it could delete a wrong or something worng due to wrong index array between posattriblist
and seed array.

As discussion with Nanne who suggests to keep a seed always be appeared by un-limited from the searching.
So I am not sure if this solution matches the idea. But it solved the bug with always a seed created by a click.

Test Plan: Compiles fine. Tests fine in F3 demo

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9604