od6.0: Fix for the mantis bug#5294, incorrect display of the well name in the well tie to seismic windows
Summary: Earlier the name of the previous well was displayed which was selected during the first instance of creation of the dialog box. This was happening because the window was non modal and was created only once. Now the multiid of the previous run and current run are compared and then decision is taken whether to create the new dialog box or raise the oolder instance

Test Plan: The change has been tested by selecting default survey name initially and then adding rest of the well in the well menu. The dialog box was relaunched from separate wells and repsective dialog box opened with the correct well name. When open directly from the menu the default well was displayed.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9637