od6.0: Added static tracking menu to 3D Horizon item
Summary: changes required to be able to add a menu item to a parent OD tree item. treeItemFactorySet() function in uiODSceneMgr could not be used since it does not contain a template menu to which items can be added, and no tree exists yet when uiDipSteerPIMgr is instanciated.

Test Plan: tested by adding unconformity tracker menu. Seems to be working. I may want to do some memory profiling to ensure everything is deleted properly.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9666