od6.0: Fixed crash report issue with writing changed data by Seis Browser (Mantis 5354).
Summary: Crash report shows that output stream is used at the end of the CBVS Translator's life. So order of destruction is important.

Test Plan: Didn;t get it crashing before this fix, and it doesn't crash after the fix either. It probably was a Lucky Linux thing (or maybe: Unlucky UNIX).

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D9696