[od6.0]OD-17 The second horizon and others created during autotracking appear as a horizontal line
Summary:
In the class uimpeman, the object seedpicker is not set SelSpec by the function:setSelSpec() if the previous seedpicker is set, because the engine.activeVolume()
is global vaiable. once we have new seedpicker the calling of setSelSpec is ignored by if ( newvolume != engine.activeVolume() ) condition.

for a new horizon we have a new seedpicker but it is not set selSpec which leads to wrong interpretation result.
So the solution is that we do always calling selSpec() in the function we do check if the selSepc is same as existing selSpec.

Test Plan: Compiles fine. Tested with Victoria together, the result is the expectation we want.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9738