od: OD-29, Quick UVQ displays a horizon in single color>
Summary: The title of the bug report is faulty. The horizon was displaed in single color if it was initially displayed at a section and does not depends upon UVQ. When the horizon was displayed in full from main 3D horizon right click menu or from righ click menu on horizon name, it was displed in single color.

Test Plan: The above bug has been fixed and the attribute values are now being dipayed when "In full" is selected.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9754