Made plugin failure put error in log.
Summary:
OD-147.
If a plugin fails to load, there is no way to figure out how that happened from the logs. This change puts it in the logs, so it can be forwarded to support.

Test Plan: Tested to load incomplete shared libraries, and they either missing libraries or missing functions were reported.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D9766