od6.0 - improvement on display speed for the option"display at section"
Summary:
This is same as the revision D9631 in od trunk.
he case study people complaints about the speed when moving(dragging) inline/cross line in multi horizons and z-slices. The issue is caused by intensive
computations of intersection lines. So the idea of this solution is only to calculate the intersection line between the dragging object ( inline, cross line, z-slices,..)
and horizon, not doing re-compute all intersection lines.

Test Plan: compiles fine. Tests fine by loading 10 horizons and z-slices. It saves lot of time when moving object in the scene

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9747