od: Removed od_batch_launcher
Summary: od_batch_launcher was used on Linux and Mac to launch batch programs and od_progress_viewer, as there was no other way to get the PID of batch program. Now we are able to get the PID from QProcess interface. Therefore od_batch_launcher is not needed anymore.

Test Plan:
Tested on Linux:
Launched several batch programs. The Progress window shows the progress log as before. Also checked the PID in the debugger. It matches with the one in the progress log.

Reviewers: nanne, kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9781