Rework of the reference counting.
Summary:
A new paradigm in that the normal operation is to inherit RefCount::Referenced, rather than using macros (hard to debug). Both the old version and the new version are still supported, but the old one has to be deprecated and our usage has to go away.

An ObsPtr<> class is added to observe objects, without referenc counting them.

Input from reviewer with regards to design/naming requested.

Test Plan: New tests added in test-program.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D9801