[od6.0] OD-145 When interpreting 3D horizon in 2D scene, the seed is in the 2D scene is minuscule
Summary:
The problem is caused by the difference between 2d and 3d views in visualization of marker size. since the size of marker in 2d view is defined as the same size in 3d view,
it seems too small in 2d view. So the solution is:

    defined the marker size as 2 times of marker size in 3d view.
    added a call back function in emhorizonpainter2d so that the painter properties can be changed by changing horizon track setup properties. let it does same as emhorizon3dpainter3d does.

Test Plan:
Compiles fine.
Tested fine below:

1. track 3D horizon in 2d view. Tested by survey: F3 demo,  SegSalt2, and FireStone_TGS_UticaShale
2. track 2D horizon in 2d view. Tested by survey: F3 demo,  USGS_BeaufortSea and Miano_2D_OMV

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9791