od6.0: Fixed problems with decimated datasets.
Summary: Jira Issue: OD-58. If the start/step of seismic datasets differ from survey sampling, they cannot be loaded. Hence obtained TrcKeyZSampling of stored dataset and limited it to requested TrcKeyZSampling. I have a proper solution for this, but I need to check some more cases. So committing this hack for the release. Will be modified later.

Test Plan: Able to load decimated and non-decimated datasets (if there is stored data for the requested sampling.) No issues found.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9819