od6.0: Fix for more than 1 treeitem selected in 2D Viewer
Summary:
When opening a 2D viewer, and tracking is in progress, both the VD and horizon were selected. See OD-204
This fix checks if there's already an item selected, if so, VD is not selected as well.

Test Plan: Start tracking, Right-click on horizon to show Parent-path. In new 2D viewer only the horizon is selected.

Reviewers: yuancheng

Reviewed By: yuancheng

Differential Revision: https://phabricator.dgbes.com:8443/D9827