Added support to checkout specific revision.
Summary:
When the release branch is made (RC1), it will still use a non-release branch of the external code (e.g. osgGeo). If the external code is updated, it those changes will be incorporated in the release branch, and the release branch is not locked down.

OD-225

Test Plan: Tested with checkout of HEAD and specific revision numbers. Both work as expected (inspected checkout manually to verify).

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9832