od6.0: Added function to set output component numbers for seis parallel reader
Summary:
when an external datapack is used (to avoid data duplication), it may have more or less components than the dataset to read.
In that case the class member components_ is insufficient to map input and output components.
Part of fix for OD4-168

Test Plan: tested from unconformity tracker, with the three options possible (with/without planarity, planarity is same or other volume).

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9841