od6.0: Seed should always snap to nearest event
Summary:
Picking seeds should not look at searchwindow and always snap to nearest event.
Reverted changes in https://phabricator.dgbes.com:8443/D9604 and temporarily set a large search window when the seed is adjusted.
Fixes OD-177

Test Plan:
Picked seeds with single clicks and with drag on minimum events while the tracker settings were set to Max.
The seeds as well as the connected line is nicely adjusted to the max event.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D9843