[od6.0] Solved (PE) "Deleting a HiddenParam prematurely" messages.
Summary:
Solved (PE) "Deleting a HiddenParam prematurely" messages. Happens if exiting
OpendTect with a seed catching object (e.g. horizon) in the scene. This was
introduced by revision r42515/D9512, in which the "tracking implementation on
random line" was ported to od6.0. I can't think of any valid reason why the
removeParam(.) statement should not be there in 2 out of the 8 cases in this
revision.

Test Plan:
No more programming errors (from this source) when exiting OpendTect with
e.g. a horizon in the scene. Textually checked revision r42515/D9512 on more
missing removeParam(.) statements.

Reviewers: satyaki

Reviewed By: satyaki

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9861