[od6.0] Crashfix when closing OD after 3-D auto-tracking (OD-206)
Summary:
Crashfix when closing OD after 3-D auto-tracking. Queues have to be removed
before the WorkManager is destroyed. The destruction of the static MPE-engine
comes afterwards, so it must remove its tracker(manager)s in an earlier stage,
because those have been adding the queues.

Test Plan:
Tested in F3_Demo with preloaded attribute, inline and new horizon in scene,
picking one seed and next do autotrack, either remove or keep the tracking
set-up window and then finally exit OD, either via menu or close button.
The programmer error telling that the tracker manager queue was not yet
removed when the WorkManager is destructed, is no longer printed after the
fix. Although OD did not and does not crash in my debug environment, the 
call stack attached to OD-206 strongly suggests that the disappearance of
the programmer error will have solved the crash as well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9865