Added new functions ot uiBaseObject
Summary:
We wish to move to a system where a uiBaseObject can have multiple widgets,
which can be layed out on a grid on a parent. This revision prepares that
by enabling these functions on the uiBaseObject.

OD-6

Test Plan:
OpendTect starts. A couple of dialog were tested, and seem to do their work. No
functional changes are expected. getNrWidgets() is make pure & virtual to make
sure that all classes are adapted.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9871