[OD6.0]OD-251 Cannot find stored data when restoring a session with a horizon displayed
Summary:
Actually it is not a bug. after appearing multi times of this message, the horizon session is restored.

After investigation, the problem is caused by a new horizon without attributes which are called ValNames in the code. 
when doing restore session it will call calcRandomPosAttrib function in which it will appear the message" Cannot find stored data" if there 
are no attributes in the horizon. for a old horizon made by od4 or early od version, it will not call the function calcRandomPosAttrib. I think it is related 
to the new design of attributes item in horizon tree item.

The solution is simple. just remove the message box. The reason are:

The message box is existed in the code since 2003. I am not sure if it still suit for current design.
Currently the message box doesn't influence the result of horizon loading.
The message meaning doesn't match "no attributes in horizon"
The message will appear a several times. it bothers and confuses user.
So my opinion is removing the message box.

Test Plan: Tests with restore horizon session without this problem any more.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9907