od: Setting the hostname for single host remote computation
Summary:
uiSingleBatchJobDispatcherPars::sjd_.remotehost_ was set/get using the full description of the host instead of either the hostname or the IP address. As a result the command could not be parsed. Also the constructor was not setting the target host from the arguments.
Part of if for OD-167

Test Plan: no more parsing error. It still does not run, but probably not because of this class

Reviewers: helene

Reviewed By: helene

Subscribers: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D9920