od: Decimated 2D horizons every 100m/400ft before gridding [OD-62]
Summary: As in summary. Also fixed a crash in Gridder if the points are too many and there is not enough memory to allocate the matrix. But we may need some limits on the number of input points for RBF gridder.

Test Plan: Gridded 2D horizons in F3. There is a pause of a few seconds in LinSolver but after that the gridding proceeds normally and gives smooth surfaces.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D9933