[OD]OD-260 When restoring a session with an auto tracked horizon displayed, it doubles the horizon data in the tree
Summary:
The problem is not only for auto track, it is a general issue when restore a session which contains horizon.
1. When doing restoring horizon session, Attribute is added duplicated in vishorizondisplay. firstly in the usepar() of vissurobj.cc, 
it will add attrib to horizon display. in the function of addAttribItem in uiVisPartServer, the attribute is added again by init tree item.
The number of attribs is used to create tree subitem.
2. to avoid the two times adding of attributes, we need firstly clean the existing attributes in vis horizon before create a new horizon tree item.

Test Plan:
Tested fine in 
1. save a new horizon into a session and restore it.
2. load an existing horizon which is created by previous od versions.
3. load an new created horizon.

Reviewers: kristofer, nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9906