od6.0: Pick Set: old Pick sets did not load.
Summary:
Most old Pick Sets do not have a TrcKey stored. Then, it will be undef, which was checked to return false, meaning the user got an error.
But, if we have the geom ID, we can simply calculate the TrcKey's trace and line number from geometry.

Test Plan: Tested with new sets and Ranojay's bag of old pick sets.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D9959