od: Made single computer remote process execution working
Summary:
launch command was badly formed. Made it use the same command building functions as in MMP, including cross-platform translation of the filepaths in the par file and not only in the command
Also made the nice level to actually be used, which was never ever the case before.
Fixes OD-167

Test Plan: tested lux-lux and lux-win. Testing further on windows is required. Progress window does not work in lux-win case and in both cases the launched remote process is not killed when closing the progress window, but this was the current state of the tool in pre-5.0 releases.

Reviewers: nanne, bert

Reviewed By: bert

Subscribers: kristofer, nanne, bert, ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D9928