od: Fix for Mantis Bug #5234 Twice display of the information message
Summary: The information message for the Sample interval being different in the selected line was displayed twice. The linesel call backer was being called from the inputsel call backer which has now been commented out. This occured while the extract wavelet.

Test Plan: The changes have been tested. Wavelet is still being extracted as it was extracted earlier and there is no change in properties of the wavelet. The message is now being flashed once.

Reviewers: ranojay, raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9533