od: OD-198, Impossible to use new empty mute function in vmb picking
Summary: In OD5.0, there was separate button to set the properties of mute function. That dialog box is now popped up as soon as mute definition is selected. But the empty mute function is gone. I have provided a bool field where if mute function is to be created then we can select yes or no.

Test Plan: The results from empty mute function are compareable to that of the OD5.0

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9831