od: OD-220 Tie Well To Seismics: checkshot correction (both automatic and editor) ignores points which are outside depth range of PWave log
Summary: The plot of the drift curve was shrinking automatically and points before and after sonic log were not recorded. This was beacause the depth to time model was not being updated.

Test Plan: The current fix keeps the plot size constant and updates as well as sorts the points

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9991