Made preload return RefMan's in place for pointers.
Summary:
We are removing old observation pointers that are not safe to use as a part of
OD-362.
Further, we are changing the observe() function to return a WeakPtr.

Test Plan:
More tests are added. The preload stuff will simply not work if it is incorrect.
Basic preload + display tested.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9998