od: Initial work on making Basemap Area flexible [OD-308]
Summary:
In order to include objects outside the survey box we need a flexible basemap area. I will add a UI to set it manually. But an obvious feature will be to set it automatically so as to include all the objects. This is similar to the 'view all' button in the 3D scene.

As per my plan every BasemapObject will give a Bounding box in terms of min and max Coord which can be used to determine the envelope for 'view all'. Therefore I added a virtual getBoundingBox fn to BaseMapObject. The implementation is only available for 2D horizons as a test case for now.

Also added a setScaleFactor function in uiGraphicsViewBase because when the user presses 'view all', we should zoom out to the full view (no zoom in).

Test Plan: Tested in F3 by setting a smaller survey range so as to get the 2D lines outside this range. Pressing the 'view all' button expands the Basemap area to include the 2D horizons.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10028