[OD]OD-326 crash report about horizon exporting into pdf3d
Summary:
It cannot be reproduced in exporting existing horizon. The case can be re-produced in exporting horizon is being created in auto tracking. 
at that moment, horizon has no texture, and horizon section tile is not completely created.

So the solution in od side:
1. check the section tile pointer.
2. export horizon even if it has no texture.
3. add interface to dgbpdf3d, allow it export horizon in single color without texture.

Test Plan: Compiles fine. Test in F3-demo and Utica works

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10012