[OD6.0]OD-369 Basemap: cube outline is wrong
Summary:
The problem is caused by the function: keepBendPoints() in ODPolygon. The corner point is filtered out by the method in that function.
I think the polygon points must be followed in a certain order or direction. For the time being, my solution to solve od-369 is that just change the code to do the
same contour tracing direction as in previous version(single thread version). it works.
Possibly we need look into the function:keepBendPoints in the future.

Test Plan:
  Tested in cube outline, horizon contour and intersection contour between horizon and z-slice. all of them working fine

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10038