[od] Adding, removing and adding again a stored random line in 3D scene provided a weird line with duplicated nodes.
Summary:
(1) Adding, removing and adding again a stored random line in 3D scene provided
a weird line with duplicated nodes. Introduced in revision r43316/D9946 by
Jaap about random line synchronization issues. Changing the random line ID is
a special case in which synchronization between the random line in /Geometry
and /visSurvey is gone and needs to be re-established. setNodePositions(.)
has been split in a public part, and a protected part without geometrical
restrictions. The latter got an "internalonly" option to solve the bug.
(2) Also removed special code for random lines with more than 50 nodes.
Testing did not show any difference with and without that code for such lines.
In case moving_.trigger() has been the problem in the past, this is solved
by a temporary NotifyStopper.

Test Plan:
(1) Tested in F3_Demo. Adding, removing and adding again the stored random
line through all wells in 3D scene works perfect again. Also the duplication
of a random line still works fine. The same holds if the base map window
is popped up as well.
(2) Used "Example random contour demo 1 -> MFS4", which has many nodes, to
conclude that is does not make a difference (anymore) if the special code
for 50+ nodes is removed.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10058