od: Fix for marker direction not displayed after restoring saved pick set [OD-294]
Summary:
https://jira.dgbes.com:8443/browse/OD-294?filter=-2

It seems the order of setting the PickSet to the PickSetDisplay, and adding the PickSetDisplay to the scene is changed in this revision 

https://github.com/OpendTect/OpendTect/commit/a0295b7edc0623b2baed0b64becb266b327b7731

So the scene is not set to the PickSetDisplay, while adding the PickSet, which is done later in addObject. So all the dip calculation fails because of no scene(Z-Stretch)

Now just calling an extra fullredraw(), after the display is added to the scene.

Test Plan:
1) Set direction on fly -> works
2) Load a PickSet with saved direction -> works.

Reviewers: nanne, bert

Reviewed By: bert

Subscribers: bert, raman

Differential Revision: https://phabricator.dgbes.com:8443/D10086