od6.0: OD-347, Copy cube within a polygon does not work if executed in batch
Summary: Copying a cube using batch process was not giving proper result when polygon selection was used. The incorrect par file was being passed. This has been corrected.

Test Plan: The cube is now being formed in the polygon subselection range. It is giving proper result in all the priority situations.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10062