od: OD-337 Some 2D horizon names are not displayed in basemap
Summary: Earlier when 2D-Horizons were displayed in basemap then sometimes their names were not dislayed. This error was because crd[0] was selected for display of the name. But in case of these horizons the crd[0] was undefined. Now a check has been placed which looks for the first encounter of the defined values and displays the name label there.

Test Plan: The fix has been tested using the data set for F3Demo and all the 2D Horizons were displaying their names correctly.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10021