[od6.0] Fix for memory issues (double delete and no delete).
Summary:
During work on OD-362 in od, these errors were found, and were thought
to be servere enough to be fixed in od6.0.

Test Plan:
Statistics window tested, and databack behaves as it should.
The uiTrcPositionDlg is only called from the spectral decomposition
attribute, and could not be called. As it is a removal of an release()
call, it is safer. Worst thing that could happen is a memory leak.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D10128