od6.0: Make menus scrollable when too large to display
Summary:
When menus become too large to fit on the screen, Qt adds an extra column. This does not work (anymore) on Windows, and becomes a bigger problem when a user has so many items that the number of columns don't fit on the screen either.
This fix sets a stylesheet to QMenu and makes the menu scrollable when it becomes larger than the screenheight.
Fixes OD-420

Test Plan: Tested on Linux and Windows. When adding enough items to the omf file, the menu has small arrowa at top/bottom of the screen to scroll.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10153